Rename the DronePort package to BeePort and update package references #83
