+++ src/main/java/de/theidler/create_mobile_packages/index/CMPBlocks.java
@@ -4,7 +4,7 @@ import com.simibubi.create.AllTags;
 import com.simibubi.create.foundation.data.SharedProperties;
 import com.tterrag.registrate.util.entry.BlockEntry;
 import de.theidler.create_mobile_packages.CreateMobilePackages;
-import de.theidler.create_mobile_packages.blocks.drone_port.DronePortBlock;
+import de.theidler.create_mobile_packages.blocks.bee_port.DronePortBlock;
 
 import static com.simibubi.create.foundation.data.ModelGen.customItemModel;
